+++ ImageSlideshow/Classes/Core/ImageSlideshowItem.swift
@@ -10,7 +10,11 @@ import UIKit
 /// Used to wrap a single slideshow item and allow zooming on it
 open class ImageSlideshowItem: UIScrollView, UIScrollViewDelegate {
 
+    /// Image view to hold the image
     open let imageView = UIImageView()
+
+    /// Activity indicator shown during image loading, when nil there won't be shown any
+    open let activityIndicator: ActivityIndicatorView?
 
     /// Input Source for the item
     open let image: InputSource
@@ -23,11 +27,16 @@ open class ImageSlideshowItem: UIScrollView, UIScrollViewDelegate {
 
     /// If set to true image is initially zoomed in
     open var zoomInInitially = false
-    
-    open var spinner: UIActivityIndicatorView
 
     fileprivate var lastFrame = CGRect.zero
     fileprivate var imageReleased = false
+    fileprivate var singleTapGestureRecognizer: UITapGestureRecognizer?
+    fileprivate var loadFailed = false {
+        didSet {
+            singleTapGestureRecognizer?.isEnabled = loadFailed
+            gestureRecognizer?.isEnabled = !loadFailed
+        }
+    }
 
     // MARK: - Life cycle
 
@@ -36,17 +45,15 @@ open class ImageSlideshowItem: UIScrollView, UIScrollViewDelegate {
         - parameter image: Input Source to load the image
         - parameter zoomEnabled: holds if it should be possible to zoom-in the image
     */
-    init(image: InputSource, zoomEnabled: Bool) {
-        
+    init(image: InputSource, zoomEnabled: Bool, activityIndicator: ActivityIndicatorView? = nil) {
         self.zoomEnabled = zoomEnabled
         self.image = image
-        self.spinner = UIActivityIndicatorView(activityIndicatorStyle: UIActivityIndicatorViewStyle.gray);
-        self.spinner.hidesWhenStopped = true;
+        self.activityIndicator = activityIndicator
+
+        super.init(frame: CGRect.null)
 
         imageView.clipsToBounds = true
         imageView.isUserInteractionEnabled = true
-        
-        super.init(frame: CGRect.null)
 
         setPictoCenter()
 
@@ -58,15 +65,20 @@ open class ImageSlideshowItem: UIScrollView, UIScrollViewDelegate {
         minimumZoomScale = 1.0
         maximumZoomScale = calculateMaximumScale()
 
+        if let activityIndicator = activityIndicator {
+            addSubview(activityIndicator.view)
+        }
+
         // tap gesture recognizer
         let tapRecognizer = UITapGestureRecognizer(target: self, action: #selector(ImageSlideshowItem.tapZoom))
         tapRecognizer.numberOfTapsRequired = 2
         imageView.addGestureRecognizer(tapRecognizer)
         gestureRecognizer = tapRecognizer
-        
-        
-        self.imageView.addSubview(self.spinner);
-        
+
+        singleTapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(retryLoadImage))
+        singleTapGestureRecognizer!.numberOfTapsRequired = 1
+        singleTapGestureRecognizer!.isEnabled = false
+        imageView.addGestureRecognizer(singleTapGestureRecognizer!)
     }
 
     required public init?(coder aDecoder: NSCoder) {
@@ -89,6 +101,8 @@ open class ImageSlideshowItem: UIScrollView, UIScrollViewDelegate {
             setPictoCenter()
         }
 
+        self.activityIndicator?.view.center = imageView.center
+
         // if self.frame was changed and zoomInInitially enabled, zoom in
         if lastFrame != frame && zoomInInitially {
             setZoomScale(maximumZoomScale, animated: false)
@@ -98,19 +112,20 @@ open class ImageSlideshowItem: UIScrollView, UIScrollViewDelegate {
 
         contentSize = imageView.frame.size
         maximumZoomScale = calculateMaximumScale()
-        
-        self.setSpinnerToCenter();
     }
 
     /// Request to load Image Source to Image View
     func loadImage() {
         if self.imageView.image == nil {
-            self.spinner.startAnimating();
+            activityIndicator?.show()
+
             imageReleased = false
             image.load(to: self.imageView) { image in
                 // set image to nil if there was a release request during the image load
                 self.imageView.image = self.imageReleased ? nil : image
-                self.spinner.stopAnimating();
+
+                self.activityIndicator?.hide()
+                self.loadFailed = image == nil
             }
         }
     }
@@ -118,6 +133,10 @@ open class ImageSlideshowItem: UIScrollView, UIScrollViewDelegate {
     func releaseImage() {
         imageReleased = true
         self.imageView.image = nil
+    }
+
+    func retryLoadImage() {
+        self.loadImage()
     }
 
     // MARK: - Image zoom & size
@@ -188,10 +207,6 @@ open class ImageSlideshowItem: UIScrollView, UIScrollViewDelegate {
         intendHorizon = intendHorizon > 0 ? intendHorizon : 0
         intendVertical = intendVertical > 0 ? intendVertical : 0
         contentInset = UIEdgeInsets(top: intendVertical, left: intendHorizon, bottom: intendVertical, right: intendHorizon)
-    }
-    
-    fileprivate func setSpinnerToCenter() {
-        self.spinner.frame = CGRect(x: screenSize().width/2 - 18 , y: self.imageView.frame.height/2 - 18, width: 35, height: 35);
     }
     
     private func isFullScreen() -> Bool {